Conflicts:
	web-module/webapp/WEB-INF/classes/data/pasae.lck
	web-module/webapp/WEB-INF/classes/data/pasae.log
	web-module/webapp/WEB-INF/classes/data/pasae.properties
	web-module/webapp/WEB-INF/classes/data/pasae.script

+++ dao-module/src/main/java/ar/edu/uai/paradigms/dao/SectorDAO.java
@@ -7,7 +7,7 @@ import java.util.Collection;
 
 public interface SectorDAO extends GenericDAO<Sector> {
 	
-	public Integer chequearDisponibilidad (long id_sector);
+	public Long chequearDisponibilidad(long id_sector);
 
 	public Collection<Asiento> asientosDeSector (long id_sector);
